gen package cleanup
