+++ cmd/generate-fix/internal/generate.go
@@ -1,4 +1,4 @@
-package gen
+package internal
 
 import (
 	"fmt"
