+++ cmd/generate-fix/internal/helpers.go
@@ -1,4 +1,4 @@
-package gen
+package internal
 
 import (
 	"os"
@@ -6,11 +6,11 @@ import (
 	"strings"
 )
 
-// GetImportPathRoot returns the root path to use in import statements.
+// getImportPathRoot returns the root path to use in import statements.
 // The root path is determined by stripping "$GOPATH/src/" from the current
 // working directory.  For example, when generating code within the QuickFIX/Go
 // source tree, the returned root path will be "github.com/quickfixgo/quickfix".
-func GetImportPathRoot() string {
+func getImportPathRoot() string {
 	pwd, err := os.Getwd()
 	if err != nil {
 		panic(err)
